backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
groq==0.9.0
pydantic==2.8.0
python-multipart==0.0.9
